# API/docs/swagger.yaml
openapi: 3.0.1
info:
  title: Citi+
  description: Documentation for Citi+ REST API.
  termsOfService: https://www.ratp.fr/gestion-de-vos-donnees-personnelles
  contact:
    email: l_laugier@etu-webschoolfactory.fr
  license:
    name: MIT License
    url: http://localhost:80/api/v1/documentation/LICENSE
  version: 1.0.0
...
servers:
- url: http://localhost:80/api/v1/
...
paths:
...
  /stations-ratp:
    get:
      tags:
      - ratp
      summary: Get all RATP stations
      operationId: StationsGET
      parameters:
      - name: limit
        in: query
        schema:
          type: string
      - name: offset
        in: query
        schema:
          type: string
      - name: outside
        in: query
        schema:
          type: string
          enum:
          - "true"
          - "false"
      - name: type
        in: query
        schema:
          type: string
          enum:
          - train
          - RER
          - métro
          - bus
      - name: line
        in: query
        schema:
          type: string
          enum:
          - A
          - B
          - "1"
          - "2"
          - "3"
          - 3bis
          - "4"
          - "5"
          - "6"
          - "7"
          - 7bis
          - "8"
          - "9"
          - "10"
          - "11"
          - "12"
          - "13"
          - "14"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ratpStation'
        500:
          description: Database error
          content: {}
  /station-ratp/{id}:
    get:
      tags:
      - ratp
      summary: Get a RATP station by ID
      operationId: StationGET
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ratpStation'
        500:
          description: Database error
          content: {}
...
components:
  schemas:
...
    ratpStation:
      type: object
      properties:
        identifiant:
            type: integer
            format: int64
        feature:
            type: object
            properties:
                type:
                    type: string
                    example: Feature
                geometry:
                    type: object
                    properties:
                        type:
                            type: string
                            example: Point
                        coordinates:
                            type: array
                            items:
                                type: array
                                items:
                                    type: number
                                    format: double
                        properties:
                            type: object
                    required:
                    - type
                    - coordinates
            required:
                - geometry
        name:
            type: string
            example: Châtelet
        address:
            type: string
            example: 25 rue des Volontaires - 75115
        coordinates:
            type: string
            example: 48.8414184427,2.30799364939
        train_outside:
            type: boolean
            default: false
        transport_type:
            type: string
            example: métro
        line:
            type: string
            example: 7bis
        line_index:
            type: integer
            format: int64    
...